Demonstration of Template Matching in OpenCv
In this mini project , we are finding waldo (a charachter ) in a given image using  the concept of template matching .There are a variety of methods to perform template matching, but in this case I am using the correlation coefficient which is specified by the flag cv2.TM_CCOEFF.

So what exactly is the cv2.matchTemplate function doing? Essentially, this function takes the image we want to find as a  “sliding window” of our waldo query image and slides it across our puzzle image from left to right and top to bottom, one pixel at a time. Then, for each of these locations, we compute the correlation coefficient to determine how “good” or “bad” the match is.